Extract `IncludeCurrentSwitch` component
