+++ frontend/src/metabase/querying/components/DatePicker/DateOperatorPicker/DateOperatorPicker.tsx
@@ -1,15 +1,9 @@
 import { useMemo } from "react";
-import { t } from "ttag";
 
-import { Group, Stack, Switch } from "metabase/ui";
+import { Group, Stack } from "metabase/ui";
 
 import { SimpleRelativeDatePicker } from "../RelativeDatePicker";
-import {
-  getIncludeCurrentLabel,
-  getIncludeCurrent,
-  setIncludeCurrent,
-} from "../RelativeDatePicker/DateIntervalPicker/utils";
-import type { DateIntervalValue } from "../RelativeDatePicker/types";
+import { IncludeCurrentSwitch } from "../RelativeDatePicker/IncludeCurrentSwitch";
 import { isIntervalValue, isRelativeValue } from "../RelativeDatePicker/utils";
 import { SimpleSpecificDatePicker } from "../SpecificDatePicker";
 import type { DatePickerOperator, DatePickerValue } from "../types";
@@ -43,26 +37,6 @@ export function DateOperatorPicker({
     }
   };
 
-  const IncludeCurrentSwitch = ({ value }: { value: DateIntervalValue }) => {
-    const includeCurrent = getIncludeCurrent(value);
-
-    const handleIncludeCurrentSwitch = () => {
-      onChange(setIncludeCurrent(value, !includeCurrent));
-    };
-
-    return (
-      <Switch
-        aria-checked={includeCurrent}
-        checked={includeCurrent}
-        data-testid="include-current-interval-option"
-        label={t`Include ${getIncludeCurrentLabel(value.unit)}`}
-        labelPosition="right"
-        onChange={handleIncludeCurrentSwitch}
-        size="sm"
-      />
-    );
-  };
-
   return (
     <Stack>
       <Group>
@@ -72,7 +46,7 @@ export function DateOperatorPicker({
         )}
       </Group>
       {isRelativeValue(value) && isIntervalValue(value) && (
-        <IncludeCurrentSwitch value={value} />
+        <IncludeCurrentSwitch value={value} onChange={onChange} />
       )}
       {value?.type === "specific" && (
         <SimpleSpecificDatePicker value={value} onChange={onChange} />
