+++ frontend/src/metabase/querying/components/DatePicker/RelativeDatePicker/IncludeCurrentSwitch/index.ts
@@ -1,0 +1,1 @@
+export * from "./IncludeCurrentSwitch";
